@@ -37,29 +37,29 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: 1.0.6
-  shared_preferences: 2.2.0
+  cupertino_icons: ^1.0.8
+  shared_preferences: ^2.2.3
   flutter_statusbarcolor_ns: 0.5.0
-  path: 1.8.3
-  get: 4.6.5
+  path: ^1.9.0
+  get: ^4.6.6
   flutter_linkify: 6.0.0
-  flutter_form_builder: 9.1.1
-  intl: 0.18.1
-  uuid: 3.0.7
-  go_router: 10.1.2
-  form_builder_validators: 9.0.0
-  form_builder_extra_fields: 10.0.0
-  animate_do: 3.0.2
+  flutter_form_builder: ^9.3.0
+  intl: ^0.19.0
+  uuid: ^4.4.0
+  go_router: ^14.2.0
+  form_builder_validators: ^10.0.1
+  form_builder_extra_fields: ^10.2.0
+  animate_do: ^3.3.4
   get_storage: 2.1.1
   cached_network_image: ^3.3.0
-  image_picker: 1.0.4
-  share_plus: 7.1.0
+  image_picker: ^1.1.2
+  share_plus: ^9.0.0
   flutter_launcher_icons: 0.13.1
-  flutter_svg: 2.0.7
-  flutter_hooks: 0.20.1
+  flutter_svg: ^2.0.10+1
+  flutter_hooks: ^0.20.5
   font_awesome_flutter: ^10.5.0
   url_launcher: ^6.1.14
-  supabase_flutter: ^1.10.18
+  supabase_flutter: ^2.5.5
   timeago: ^3.5.0
   flutter_native_splash: ^2.3.5
   timeago_flutter: ^3.5.0
@@ -74,7 +74,6 @@ dependencies:
   faker: ^2.1.0
   shimmer: ^3.0.0
   accordion: ^2.6.0
-
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -84,7 +83,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^4.0.0
   json_serializable: ^6.7.1
 
 # For information on the generic Dart part of this file, see the
