@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class CallButton extends StatelessWidget {
-  const CallButton({Key? key, required this.uri}) : super(key: key);
+  const CallButton({super.key, required this.uri});
 
   final Uri uri;
 
@@ -10,7 +10,7 @@ class CallButton extends StatelessWidget {
   Widget build(BuildContext context) {
     return IconButton(
         style: ButtonStyle(
-            backgroundColor: MaterialStatePropertyAll(Colors.green.shade600)),
+            backgroundColor: WidgetStatePropertyAll(Colors.green.shade600)),
         onPressed: () {
           launchUrl(uri);
         },
